The fix in r28/r29 appears to be causing some boss debuffs not to show up, probably because they are incorrectly labeled by the game as player cast debuffs.  For example, Convulsive Shadows on Iron Maidens and Retched Blackrock on Oregorger.  Thanks m0rph0 for the help testing.
This version uses a different fix to the Exhaustion debuff (Ticket #6), by
only ignoring those specific player debuffs that are causing the issue.
GridStatusRaidDebuff now also checks the ignore_ids before displaying a
debuff.  Exhaustion and Weakened Soul have been added to the ignore_ids.
With this solution, any time a raid is released with a debuff with the same
name as a player debuff, the player debuff will need to be added to the
ignore_ids.  Unfortunately, the general solution bug fix had side effects.